refactor: Remove pkg/find/vulnerabilities package
