--- pkg/find/vulnerabilities/scanner_test.go
@@ -1,1 +1,0 @@
-package vulnerabilities_test
