+++ pkg/vulnerabilityreport/builder_test.go
@@ -1,6 +1,8 @@
 package vulnerabilityreport_test
 
 import (
+	"testing"
+
 	"github.com/aquasecurity/starboard/pkg/apis/aquasecurity/v1alpha1"
 	"github.com/aquasecurity/starboard/pkg/vulnerabilityreport"
 	"github.com/stretchr/testify/assert"
@@ -8,8 +10,6 @@ import (
 	appsv1 "k8s.io/api/apps/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes/scheme"
-
-	"testing"
 )
 
 func TestBuilder(t *testing.T) {
