+++ pkg/vulnerabilityreport/plugin.go
@@ -1,0 +1,28 @@
+package vulnerabilityreport
+
+import (
+	"io"
+
+	"github.com/aquasecurity/starboard/pkg/apis/aquasecurity/v1alpha1"
+	"github.com/aquasecurity/starboard/pkg/docker"
+	corev1 "k8s.io/api/core/v1"
+)
+
+// Plugin defines the interface between Starboard and static vulnerability
+// scanners.
+type Plugin interface {
+
+	// GetScanJobSpec describes the pod that will be created by Starboard when
+	// it schedules a Kubernetes job to scan the workload with the specified
+	// descriptor.
+	// The second argument maps container names to Docker registry credentials,
+	// which can be passed to the scanner as environment variables with values
+	// set from secrets.
+	GetScanJobSpec(spec corev1.PodSpec, credentials map[string]docker.Auth) (
+		corev1.PodSpec, []*corev1.Secret, error)
+
+	// ParseVulnerabilityScanResult is a callback to parse and convert logs of
+	// the pod controlled by the scan job to the Starboard model.
+	ParseVulnerabilityScanResult(imageRef string, logsReader io.ReadCloser) (
+		v1alpha1.VulnerabilityScanResult, error)
+}
