+++ pkg/vulnerabilityreport/plugin_test.go
@@ -1,0 +1,1 @@
+package vulnerabilityreport_test
