+++ pkg/vulnerabilityreport/scanner_test.go
@@ -1,0 +1,1 @@
+package vulnerabilityreport_test
